third_party/blink/web_tests/platform/generic/fast/events/page-visibility-iframe-move-test-expected.txt

This test checks that an iframe that moves between pages with different visibility will have the correct visibility value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Window 1 Loaded
Window 2 Loaded
PASS window.document.hidden is false
PASS window2.document.hidden is false
PASS iframe.contentDocument.hidden is false
PASS window.document.hidden is true
PASS window2.document.hidden is false
PASS iframe.contentDocument.hidden is false
Adopted iframe to Window 1
PASS window.document.hidden is true
PASS window2.document.hidden is false
PASS iframe.contentDocument.hidden is true
PASS successfullyParsed is true

TEST COMPLETE
